As discussed in https://github.com/NUS-HPC-AI-Lab/OpenDiT/issues/61. 

Stuff done so far:

* Added support for pushing the final experimental files from `train.py` to the Hub if `push_to_hub` flag is enabled.  
* Loading from the Hub (only for the image models).
* Incorporate support for using `accelerate` to massively speed up model loading time. 

You can use the script below to verify the third point:

```python
from opendit.models.dit import DiT_models
from accelerate import init_empty_weights
from accelerate.utils import set_module_tensor_to_device
from huggingface_hub import hf_hub_download
from contextlib import nullcontext
import torch
import argparse


def main(args):
    init_ctx = init_empty_weights if args.speedup_loading else nullcontext
    image_size = 256
    input_size = image_size // 8
    dtype = torch.float32

    ckpt_path = hf_hub_download(repo_id="sayakpaul/OpenDiT", filename="OpenDiT-DiT-XL-2-256x256.pt")
    state_dict = torch.load(ckpt_path, map_location="cpu")

    with init_ctx():
        model = DiT_models["DiT-XL/2"](
            use_video=False,
            input_size=input_size,
            num_classes=1000,
            enable_flashattn=False,
            enable_layernorm_kernel=False,
            dtype=dtype,
            text_encoder=None,
        )
    if args.speedup_loading:
        for param_name, param in state_dict.items():
            set_module_tensor_to_device(model, param_name, "cpu", value=param)
    else:
        model.load_state_dict(state_dict)


if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument("--speedup_loading", action="store_true")
    args = parser.parse_args()

    main(args)
```

Here are the results:

**_Without `accelerate`_**: 

```bash
CPU times: user 236 ms, sys: 34.3 ms, total: 270 ms
Wall time: 34.3 s
```

**_With `accelerate`_**:

```bash
CPU times: user 61.7 ms, sys: 9.31 ms, total: 71 ms
Wall time: 8.14 s
```

While comparing the results, I ensured that the model was cached and wasn't re-downloaded. 

I have also taken the liberty to host the OpenDiT checkpoint here: https://huggingface.co/sayakpaul/OpenDiT/. Would be nice if you could create an org on the Hub so that we can transfer this checkpoint to your org (or you can do so yourself if it's okay).  

I want to get a review first from the library maintainers to check if they're okay with the changes. I can then propagate them to the video scripts too. Will add a section in the README about the usage as well. 

Cc: @oahzxl

Cc: @LysandreJik for awareness.