@sayakpaul 2024-07-19
May I know why this was closed without any reasoning and despite having agreement on having the feature? I find that to be outright disrespectful.