@oahzxl 2024-07-19
we are very thankful for your contribution. the reason why we close this pr is that it is already outdated because we dont use that model any more. and currently all latest checkpoints are on huggingface.