@sayakpaul 2024-07-19
Thanks. 

Can the checkpoints be directly loaded into the model without manual downloads?