@oahzxl 2024-07-19
yeah currently we use model.from_pretrained to get models from huggingface